removed cluster utils and migrated makeHostName
