+++ lib/cluster/services/messaging.js
@@ -1,7 +1,5 @@
 'use strict';
 
-var makeHostName = require('../../utils/cluster').makeHostName;
-
 var networkMapping = {
     'network:connect': 'connect',
     'network:disconnect': 'disconnect',
@@ -86,6 +84,29 @@ function makeConfigurations(context) {
     config.assignment = env.assignment;
 
     return config;
+}
+
+function makeHostName(hostname, port, nameSpace) {
+    var name;
+
+    if (!hostname.match(/http/)) {
+        hostname = 'http://' + hostname;
+    }
+
+    var lastChar = hostname[hostname.length - 1];
+
+    if (lastChar !== ':') {
+        name = hostname + ':' + port;
+    }
+    else {
+        name = hostname + port;
+    }
+
+    if (nameSpace) {
+        return name + '/' + nameSpace
+    }
+
+    return name;
 }
 
 module.exports = function messaging(context, logger) {
